@@ -14,18 +14,20 @@ description: Scanbot Barcode Scanner SDK Example App for Flutter
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
+  flutter: ^2.0.1
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Scanbot Barcode Scanner SDK
-  barcode_scanner: ^1.1.0
+  barcode_scanner:
+    path: ../scanbot-barcode-sdk-flutter/
 
-  path_provider: ^1.6.27
-  image_picker: 0.6.3+1
-  cupertino_icons: ^0.1.2
+  path_provider: ^2.0.1
+  image_picker: ^0.7.2
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
