@@ -6,10 +6,10 @@ import 'package:flutter/rendering.dart';
 class MenuItemWidget extends StatelessWidget {
   final String title;
   final GestureTapCallback onTap;
-  final IconData startIcon;
-  final IconData endIcon;
+  final IconData? startIcon;
+  final IconData? endIcon;
 
-  MenuItemWidget(this.title, {this.onTap, this.startIcon, this.endIcon});
+  MenuItemWidget(this.title, {required this.onTap,  this.startIcon,  this.endIcon});
 
   @override
   Widget build(BuildContext context) {
